+++ src/lib/pages/code-details/index.tsx
@@ -95,7 +95,7 @@ const CodeDetailsBody = observer(
             title="Public Code Description"
             description={publicProject.publicCodeData.description}
             textLine={2}
-            icon={<CustomIcon name="public-project" ml="0" my="0" />}
+            icon={<CustomIcon name="website" ml="0" my="0" />}
           />
         )}
         <Divider borderColor="pebble.700" my={12} />